requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Mail
Flask-WTF
Werkzeug
WTForms==2.2.*
email-validator
bootstrap-flask
python-dateutil